Moving a lot of reusable code to server_functions.ts
A lot of functions located in API routers were being needed in other places. So I just moved all server function code to this file and have the routers call the functions they need.
Also functions for the Defense MO's
And the rest of the files I dont wan to make extra commits for.