Fixed some issues
There was an issue showing stats over time. This seems related to the ints for the values getting overflowed.

Also stopped the planet images from being refreshed on every subsequent API request if we cant find an appropriate image.

Fixed issue with the patchnotes. Encountering an invalid character during the bbcode parse